requirements.txt
Django>=5.1.3
django-allauth>=0.65.2
python-dotenv>=1.0.1
requests>=2.32.3
PyJWT>=2.10.1
cryptography>=44.0.0
